@@ -1,18 +1,18 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net7.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Zamin.Core.ApplicationServices" Version="6.0.1" />
-  </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="Zamin.Core.ApplicationServices" Version="7.0.0-beta.1" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\ZaminTemplate.Core.Contracts\ZaminTemplate.Core.Contracts.csproj" />
-    <ProjectReference Include="..\ZaminTemplate.Core.Domain\ZaminTemplate.Core.Domain.csproj" />
-  </ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\ZaminTemplate.Core.Contracts\ZaminTemplate.Core.Contracts.csproj" />
+		<ProjectReference Include="..\ZaminTemplate.Core.Domain\ZaminTemplate.Core.Domain.csproj" />
+	</ItemGroup>
 
 </Project>
