@@ -1,14 +1,14 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net7.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Zamin.Core.Domain" Version="6.0.1" />
-    <PackageReference Include="Zamin.Core.Domain.Toolkits" Version="6.0.1" />
-  </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="Zamin.Core.Domain" Version="7.0.0-beta.1" />
+		<PackageReference Include="Zamin.Core.Domain.Toolkits" Version="7.0.0-beta.1" />
+	</ItemGroup>
 
 </Project>
