@@ -1,13 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Zamin.Infra.Data.Sql.Queries" Version="6.0.1" />
+    <PackageReference Include="Zamin.Infra.Data.Sql.Queries" Version="7.0.0-beta.1" />
   </ItemGroup>
 
   <ItemGroup>
