@@ -1,17 +1,17 @@
 ﻿using Zamin.Core.Domain.Events;
 
-namespace ZaminTemplate.Core.Domain.Blogs.Events
+namespace ZaminTemplate.Core.Domain.Blogs.Events;
+
+public class BlogCreated : IDomainEvent
 {
-    public class BlogCreated:IDomainEvent
+    public string BusinessId { get; private set; }
+    public string Title { get; private set; }
+    public string Description { get; private set; }
+
+    public BlogCreated(string businessId, string title, string description)
     {
-        public string BusinessId { get;private set; }
-        public string Title { get; private set; }
-        public string Description { get; private set; }
-        public BlogCreated(string businessId, string title,string description)
-        {
-            BusinessId = businessId;
-            Title = title;
-            Description = description;
-        }
+        BusinessId = businessId;
+        Title = title;
+        Description = description;
     }
 }
