@@ -1,16 +1,14 @@
-﻿using ZaminTemplate.Core.Contracts.Blogs.CommandRepositories;
+﻿using Zamin.Infra.Data.Sql.Commands;
+using ZaminTemplate.Core.Contracts.Blogs.Commands;
 using ZaminTemplate.Core.Domain.Blogs.Entities;
 using ZaminTemplate.Infra.Data.Sql.Commands.Common;
-using Zamin.Infra.Data.Sql.Commands;
 
-namespace ZaminTemplate.Infra.Data.Sql.Commands.Blogs
+namespace ZaminTemplate.Infra.Data.Sql.Commands.Blogs;
+
+public class BlogCommandRepository
+    : BaseCommandRepository<Blog, ZaminTemplateCommandDbContext>, IBlogCommandRepository
 {
-    public class BlogCommandRepository : 
-        BaseCommandRepository<Blog, ZaminTemplateCommandDbContext>, 
-        IBlogCommandRepository
+    public BlogCommandRepository(ZaminTemplateCommandDbContext dbContext) : base(dbContext)
     {
-        public BlogCommandRepository(ZaminTemplateCommandDbContext dbContext) : base(dbContext)
-        {
-        }
     }
 }
