@@ -1,21 +1,22 @@
 ﻿using Microsoft.EntityFrameworkCore;
-using ZaminTemplate.Core.Domain.Blogs.Entities;
-using Zamin.Core.Domain.Toolkits.ValueObjects;
+using System.Reflection;
+using Zamin.Extensions.Events.Outbox.Dal.EF;
 using Zamin.Infra.Data.Sql.Commands;
+using ZaminTemplate.Core.Domain.Blogs.Entities;
 
-namespace ZaminTemplate.Infra.Data.Sql.Commands.Common
+namespace ZaminTemplate.Infra.Data.Sql.Commands.Common;
+
+public class ZaminTemplateCommandDbContext : BaseOutboxCommandDbContext
 {
-    public class ZaminTemplateCommandDbContext : BaseCommandDbContext
-    {
-        public DbSet<Blog> Blogs { get; set; }
-        public ZaminTemplateCommandDbContext(DbContextOptions<ZaminTemplateCommandDbContext> options) : base(options)
-        {
-        }
-        protected override void ConfigureConventions(ModelConfigurationBuilder configurationBuilder)
-        {
+    public DbSet<Blog> Blogs { get; set; }
 
-            configurationBuilder.Properties<Description>().HaveConversion<DescriptionConversion>();
-            configurationBuilder.Properties<Title>().HaveConversion<TitleConversion>();
-        }
+    public ZaminTemplateCommandDbContext(DbContextOptions<ZaminTemplateCommandDbContext> options) : base(options)
+    {
+    }
+
+    protected override void OnModelCreating(ModelBuilder builder)
+    {
+        builder.ApplyConfigurationsFromAssembly(Assembly.GetExecutingAssembly());
+        base.OnModelCreating(builder);
     }
 }
