@@ -1,12 +1,11 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Zamin.Infra.Data.Sql.Queries;
 
-namespace ZaminTemplate.Infra.Data.Sql.Queries.Common
+namespace ZaminTemplate.Infra.Data.Sql.Queries.Common;
+
+public class ZaminTemplateQueryDbContext : BaseQueryDbContext
 {
-    public class ZaminTemplateQueryDbContext : BaseQueryDbContext
+    public ZaminTemplateQueryDbContext(DbContextOptions options) : base(options)
     {
-        public ZaminTemplateQueryDbContext(DbContextOptions options) : base(options)
-        {
-        }
     }
 }
